os/bluestore/bluestore_types: uint64_t for ref_map

+++ src/os/bluestore/bluestore_types.h
@@ -151,7 +151,7 @@ struct bluestore_extent_ref_map_t {
   struct record_t {
     uint32_t length;
     uint32_t refs;
-    record_t(uint32_t l=0, uint32_t r=0) : length(l), refs(r) {}
+    record_t(uint64_t l=0, uint32_t r=0) : length(l), refs(r) {}
     void encode(bufferlist& bl) const {
       small_encode_varint_lowz(length, bl);
       small_encode_varint(refs, bl);
@@ -163,10 +163,10 @@ struct bluestore_extent_ref_map_t {
   };
   WRITE_CLASS_ENCODER(record_t)
 
-  map<uint32_t,record_t> ref_map;
+  map<uint64_t,record_t> ref_map;
 
   void _check() const;
-  void _maybe_merge_left(map<uint32_t,record_t>::iterator& p);
+  void _maybe_merge_left(map<uint64_t,record_t>::iterator& p);
 
   void clear() {
     ref_map.clear();
@@ -175,11 +175,11 @@ struct bluestore_extent_ref_map_t {
     return ref_map.empty();
   }
 
-  void get(uint32_t offset, uint32_t len);
-  void put(uint32_t offset, uint32_t len, vector<bluestore_pextent_t> *release);
+  void get(uint64_t offset, uint32_t len);
+  void put(uint64_t offset, uint32_t len, vector<bluestore_pextent_t> *release);
 
-  bool contains(uint32_t offset, uint32_t len) const;
-  bool intersects(uint32_t offset, uint32_t len) const;
+  bool contains(uint64_t offset, uint32_t len) const;
+  bool intersects(uint64_t offset, uint32_t len) const;
 
   void encode(bufferlist& bl) const;
   void decode(bufferlist::iterator& p);
